chapter2-extracting-data-from-pdfs/extracting-pdf-images/requirements.txt
PyMuPDF
Pillow